feat: move interface out of submodule declarations and add simplemap as a submodule
